Start of restructuring modules. Profile and standings converted
