+++ app/modules/acquisitions_display.js
