+++ app/modules/all_acquisitions.js
