+++ app/modules/basketball_standings_router_template.js
@@ -10,10 +10,10 @@ var assert = require('assert');
 
 module.exports = function(req, res, db, year, playoffs) {
 
-	var sport = 'football';
+	var sport = "basketball";
 
-	// url for football standings
-	var url = 'http://games.espn.com/ffl/standings?leagueId=154802&seasonId=' + year;
+	// url for basketball standings
+	var url = 'http://games.espn.com/fba/standings?leagueId=100660&seasonId=' + year;
 
 	request(url, function(error, response, html) {
 
@@ -27,12 +27,12 @@ module.exports = function(req, res, db, year, playoffs) {
 			var division, team, wins, losses, ties, win_per
 			var h2h_standings = new Array();
 
-			var PF, PA
-			var other_standings = new Array();
+			var h2h_rank, team, FG_PCT, FT_PCT, THREEPM, REB, AST, STL, BLK, TO, PTS
+			var roto_standings = new Array();
 
 			// scraping h2h records and standings
 			// for each team row in the h2h standings
-			$("tr[class=tableBody]").each(function(i, element) {
+			$('tr[class=tableBody]').each(function(i, element) {
 
 				// store sraped data for each team in json
 				var json1 = {
@@ -52,7 +52,6 @@ module.exports = function(req, res, db, year, playoffs) {
 				losses = wins.next();
 				ties = losses.next();
 				win_per = ties.next();
-				
 
 				// inserting scraped data converting either to TEXT, INT, or FLOAT
 				json1.division = division.text();
@@ -62,54 +61,63 @@ module.exports = function(req, res, db, year, playoffs) {
 				json1.ties = parseInt(ties.text());
 				json1.win_per = parseFloat(win_per.text()).toFixed(3);
 
-
 				// push each team's json of data into array of all teams
 				h2h_standings.push(json1);
-				
-			}); // end of each h2h row
-			//console.log(h2h_standings);
+			})
 
-			// scrape for PF and PA for every other team
-			$('tr.evenRow.bodyCopy.sortableRow').each(function(i, element) {
+			// scraping roto stats and standings
+			// for each team row in roto standings (note space " " in class name; it's possible)
+			$('tr[class="tableBody sortableRow"]').each(function(i, element) {
 
+				// store scraped data for each team as json
 				var json2 = {
+					h2h_rank: "",
 					team: "",
-					PF: "",
-					PA: ""
-				};
+					FG_PCT: "",
+					FT_PCT: "",
+					THREEPM: "",
+					REB: "",
+					AST: "",
+					STL: "",
+					BLK: "",
+					TO: "",
+					PTS: ""
+				}
 
-				team = $(this).children().children();
-				PF = team.parent().next();
-				PA = PF.next();
+ 				// traversing DOM
+ 				h2h_rank = $(this).children().first();
+ 				// hitting categories
+ 				team = h2h_rank.next();
+ 				FG_PCT = team.next().next();
+ 				FT_PCT = FG_PCT.next();
+ 				THREEPM = FT_PCT.next();
+ 				REB = THREEPM.next();
+ 				AST = REB.next();
+ 				STL = AST.next();
+ 				BLK = STL.next();
+ 				TO = BLK.next();
+ 				PTS = TO.next();
+ 
+ 				// store each scraped value in json as TEXT, INT, or FLOAT
+ 				json2.h2h_rank = parseInt(h2h_rank.text());
+ 				json2.team = team.text();
+ 				json2.FG_PCT = parseFloat(FG_PCT.text()).toFixed(4);
+ 				json2.FT_PCT = parseFloat(FT_PCT.text()).toFixed(4);
+ 				json2.THREEPM = parseInt(THREEPM.text());
+ 				json2.REB = parseInt(REB.text());
+ 				json2.AST = parseInt(AST.text());
+ 				json2.STL = parseInt(STL.text());
+ 				json2.BLK = parseInt(BLK.text());
+ 				json2.TO = parseInt(TO.text());
+ 				json2.PTS = parseInt(PTS.text());
+ 
+ 				// push each team's json of data into array of all teams
+ 				roto_standings.push(json2);
 
-				json2.team = team.text();
-				json2.PF = parseFloat(PF.text()).toFixed(1);
-				json2.PA = parseFloat(PA.text()).toFixed(1);
-
-				other_standings.push(json2);
-			}); // end of each even row
-
-			// scrape for PF and PA for every other team
-			$('tr.oddRow.bodyCopy.sortableRow').each(function(i, element) {
-
-				var json3 = {
-					team: "",
-					PF: "",
-					PA: ""
-				};
-
-				team = $(this).children().children();
-				PF = team.parent().next();
-				PA = PF.next();
-
-				json3.team = team.text();
-				json3.PF = parseFloat(PF.text()).toFixed(1);
-				json3.PA = parseFloat(PA.text()).toFixed(1);
-
-				other_standings.push(json3);
-			}); // end of each odd row
-
-			//console.log(other_standings);
+			})
+			
+			//console.log(standings);
+			//console.log(roto_standings);
 
 			///// DEFINE FUNCTIONS (so far still only defined inside the request) ///// 
 
@@ -118,8 +126,8 @@ module.exports = function(req, res, db, year, playoffs) {
 			var insertDocument = function(db, year, callback) {
 
 				// set collections
-				var collection1 = db.collection('football_h2h_' + year);
-				var collection2 = db.collection('football_roto_' + year);
+				var collection1 = db.collection('basketball_h2h_' + year);
+				var collection2 = db.collection('basketball_roto_' + year);
 
 				// remove all documents from collections to start fresh
 				collection1.remove({});
@@ -131,7 +139,7 @@ module.exports = function(req, res, db, year, playoffs) {
 				collection1.insert(h2h_standings, function(err, result) {
 
 			    	// insert roto standings array
-					collection2.insert(other_standings, function(err, result) {
+					collection2.insert(roto_standings, function(err, result) {
 			    		
 			    		// assert to make sure no error
 				    	assert.equal(err, null);
@@ -150,6 +158,8 @@ module.exports = function(req, res, db, year, playoffs) {
 				});
 			}
 
+
+
 ///// EXECUTE SCRIPT /////
 
 			// call insertDocumet asynchronously, but don't use db from callback as we need to use db from argument to find and get from to render
@@ -157,22 +167,24 @@ module.exports = function(req, res, db, year, playoffs) {
 
 				console.log("All documents uploaded");
 
+				// initialize year as argument for python script
 				var options = {
 					args: [year]
 				}
 
 				// run standings.py from python-shell to update collections with roto and trifecta points
-				pyshell.run("football_standings.py", options, function(err) {
+				pyshell.run("basketball_standings.py", options, function(err) {
 					
 					if (err) throw err;
 					console.log("Python script complete");
 
 					// initialize display database queries
 					var disp_h2h_standings = null;
+					var disp_roto_standings = null;
 					var disp_trifecta_standings = null;
 
 					// pull from mongodb and display new data after python script finishes
-					db.collection('football_h2h_' + year).find({}, {"_id": 0}).sort({"trifecta_points": -1}).toArray(function(e, docs) {
+					db.collection('basketball_h2h_' + year).find({}, {"_id": 0}).sort({"win_per": -1}).toArray(function(e, docs) {
 						//console.log(docs);
 						console.log("Displaying h2h data...")
 						disp_h2h_standings = docs;
@@ -180,21 +192,29 @@ module.exports = function(req, res, db, year, playoffs) {
 						complete();
 					});
 
-					// if playoffs are true (playoffs are done)
+					db.collection('basketball_roto_' + year).find({}, {"_id": 0}).sort({"roto_trifecta_points": -1}).toArray(function(e, docs) {
+						//console.log(docs);
+						console.log("Displaying roto data...")
+						disp_roto_standings = docs;
+						// call complete to see if both finds are done
+						complete();
+					});
+
+					// if playoffs are true (playoffs started)
 					if (playoffs === true) {
-						// initialize year as argument for python script
-						var options = {
-							args: [year]
-						};
 
 						var playoff_scrape = require('./playoffs_router_template.js')(req, res, db, sport, year, function(err, call) {
 							
+							var options = {
+								args: [year]
+							};
+							
 							// run python script to initialize trifecta database
-							pyshell.run('football_playoffs.py', options, function(err) {
+							pyshell.run('basketball_playoffs.py', options, function(err) {
 								if (err) throw err;
 								console.log('Playoff python script complete');
 
-								db.collection('football_trifecta_' + year).find({}, {"_id": 0}).sort({"total_trifecta_points": -1}).toArray(function(e, docs) {
+								db.collection('basketball_trifecta_' + year).find({}, {"_id": 0}).sort({"total_trifecta_points": -1}).toArray(function(e, docs) {
 									//console.log(docs);
 									console.log("Displaying playoff data...");
 									disp_trifecta_standings = docs;
@@ -208,11 +228,12 @@ module.exports = function(req, res, db, year, playoffs) {
 					var complete = function() {
 
 						if (playoffs === true) {
-							if (disp_h2h_standings !== null && disp_trifecta_standings !== null) {
+							if ((disp_h2h_standings !== null && disp_roto_standings !== null) && disp_trifecta_standings !== null) {
 
 								// render to baseball_standings
-								res.render('football_standings', {
+								res.render('basketball_standings', {
 									h2h_standings: disp_h2h_standings,
+									roto_standings: disp_roto_standings,
 									trifecta_standings: disp_trifecta_standings,
 									year: year,
 									playoffs: playoffs
@@ -220,21 +241,23 @@ module.exports = function(req, res, db, year, playoffs) {
 							}
 						}
 						else {
-							if (disp_h2h_standings !== null) {
+							if (disp_h2h_standings !== null && disp_roto_standings !== null) {
 
 								// render to baseball_standings
-								res.render('football_standings', {
+								res.render('basketball_standings', {
 									h2h_standings: disp_h2h_standings,
+									roto_standings: disp_roto_standings,
 									year: year,
 									playoffs: playoffs
 								});
 							}
 						}
-					} // end of complete function
+					}				
 
-				}); // end of pyshell
-			}); // end of insertDocument
+				}); // end of pyshell 
+			}); // end of insertDocument 
 
 		} // end of if(!error)
-	}) // end of request
-} // end of module export
+	}) // end of request		
+
+}
