+++ app/modules/football_acquisitions_aux.js
