+++ app/modules/stats_display.js
